Updated svelte preprocess config & moved tests to align better with sveltekit file naming practices.
