+++ svelte.config.js
@@ -5,7 +5,24 @@ import adapter from '@sveltejs/adapter-node';
 const config = {
   // Consult https://github.com/sveltejs/svelte-preprocess
   // for more information about preprocessors
-  preprocess: preprocess(),
+  preprocess: preprocess({
+    babel: {
+      presets: [
+        [
+          '@babel/preset-env',
+          {
+            loose: true,
+            // No need for babel to resolve modules
+            modules: false,
+            targets: {
+              // ! Very important. Target es6+
+              esmodules: true,
+            },
+          },
+        ],
+      ],
+    },
+  }),
 
   kit: {
     adapter: adapter({
